user table v1

+++ app/actions/user-actions.ts
@@ -1,6 +1,8 @@
 "use server";
 
 import prisma from "@/lib/prisma";
+import { revalidatePath } from "next/cache";
+import { auth } from "@/lib/auth";
 
 // fetch all the users
 export async function fetchUsers(page = 1, pageSize = 10) {
@@ -49,17 +51,67 @@ export async function fetchUsers(page = 1, pageSize = 10) {
 
 // delete a user
 export async function deleteUser(userId: string) {
+  const session = await auth();
+
+  if (session?.user.role !== 1) {
+    return {
+      success: false,
+      msg: "Unauthorized Action",
+    };
+  }
+
   try {
     await prisma.user.delete({
       where: {
         id: userId,
       },
     });
+
+    revalidatePath("/dashboard/users");
+
+    return {
+      success: true,
+      msg: "User deleted successfully",
+    };
   } catch (err) {
     console.log(err);
     return {
       success: false,
       msg: "Failed to delete user",
+    };
+  }
+}
+
+// delete users
+export async function deleteUsers(userIds: string[]) {
+  const session = await auth();
+
+  if (session?.user.role !== 1) {
+    return {
+      success: false,
+      msg: "Unauthorized Action",
+    };
+  }
+  try {
+    await prisma.user.deleteMany({
+      where: {
+        id: {
+          in: userIds,
+        },
+      },
+    });
+
+    revalidatePath("/dashboard/users");
+
+    return {
+      success: true,
+      msg: "Users deleted successfully",
+    };
+  } catch (err) {
+    console.log(err);
+    return {
+      success: false,
+      msg: "Failed to delete users",
     };
   }
 }
